% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{euclidean}
\alias{euclidean}
\title{Compute euclidean distance}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{a vector of values}

\item{b}{a vector of values}
}
\description{
Return the euclidean distance between two points
}
