% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{explain_code_safe}
\alias{explain_code_safe}
\title{Safe version of explain from the ICD package}
\usage{
explain_code_safe(code)
}
\arguments{
\item{code}{An icd-9 or icd-10 code}
}
\description{
Return the name of code of NA
}
