% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_functions.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters}
\usage{
extract_parameters(data, var_name = "n_visits", method, week_effect = F)
}
\arguments{
\item{data}{a dataset of visit counts}

\item{var_name}{name of variable containing visit counts (or frequencies) to
fit trends to}

\item{method}{method used to fit trends}

\item{week_effect}{a logical indicator of whether to include weekly periodicity}
}
\description{
Extract model fit parameters for tends of visit codes prior to diagnosis
}
