% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_functions.R
\name{fit_trends}
\alias{fit_trends}
\title{Fit trends}
\usage{
fit_trends(count_data, param_data, method = "lm", n_fits = 500)
}
\arguments{
\item{count_data}{a dataset of visit counts for each code (and scaled accordingly)}

\item{param_data}{a dataset containing codes, names and counts to run fits over}

\item{method}{method used to fit trends}

\item{n_fits}{number of codes to fit trends to}
}
\description{
Fit trends to each of the visit count codes
}
