% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{plot_cluster_code_group}
\alias{plot_cluster_code_group}
\title{Plot cluster}
\usage{
plot_cluster_code_group(
  clust_data,
  search_code,
  size = 9,
  code_type,
  span = 0.75,
  bin_by = 1
)
}
\arguments{
\item{clust_data}{a dataset of cluster results (generated from `find_clusters()`)}

\item{search_code}{a code to search clusters for and identify a focal cluster}

\item{size}{the number of codes to include in the plot (default is 9)}

\item{code_type}{type of code to extract trends from}

\item{span}{smoothing parameter for loess fit in plot}

\item{bin_by}{number of days to bin counts by (default is 1)}
}
\description{
Plot trends from a group of codes in a given cluster
}
