% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{search_clust_terms}
\alias{search_clust_terms}
\title{Search cluster terms}
\usage{
search_clust_terms(clust_data, search_term)
}
\arguments{
\item{clust_data}{a dataset of cluster results (generated from `find_clusters()`)}

\item{search_term}{a term to search for}
}
\description{
Search clusters for a specific term in the codes used to cluster on
}
