% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{search_clusters}
\alias{search_clusters}
\title{Search clusters}
\usage{
search_clusters(clust_data, search_code)
}
\arguments{
\item{clust_data}{a dataset of cluster results (generated from `find_clusters()`)}

\item{search_code}{a code to search clusters for}
}
\description{
Search clusters for a specific code and return the focal cluster containing the code
}
