% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_dx_counts}
\alias{synth_dx_counts}
\title{Synthetic counts of visits for top 500 diagnosis codes before TB}
\format{
A data frame with 182,500 rows and 3 columns:
\describe{
  \item{dx}{ICD-9-CM code}
  \item{days_since_index}{Days Since the Index TB diagnosis}
  \item{n}{Number of patients with a visit for the given dx code}
}
}
\usage{
synth_dx_counts
}
\description{
This dataset contains synthetic data of visits for each of the top 500 diagnosis
codes before a tuberculosis diagnosis. These counts are synthetic but were
generated to match patterns consistent with patterns in claims-based datasets.
}
\keyword{datasets}
